+++ src/lib.rs
@@ -1,6 +1,7 @@
 pub mod database;
 pub mod extractors;
 pub mod measures;
+pub mod python;
 pub mod search;
 
 pub use database::{Database, HashDb};