+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# simstring-rust
+
+## v0.3.0
+
+### Features
+
+- feat: python bindings for project
+
+### Other
+
+- chore(ci): Added unit tests for python bindings
+- chore(ci): Added job to run examples for the project
+- chore(bindings): Added an example for the python example
+
 ## v0.2.0
 
 ### Features
@@ -12,7 +26,7 @@
 
 ### Other
 
-- added order indipendence test (#12)
+- added order independence test (#12)
 
 ## v0.1.2